AccountManagementService Refactored
