+++ java services/server/server/src/main/java/com/CSCI152/team4/server/Accounts/Services/AccountManagementService.java
@@ -1,8 +1,10 @@
 package com.CSCI152.team4.server.Accounts.Services;
 
 import com.CSCI152.team4.server.Accounts.Classes.*;
-import com.CSCI152.team4.server.Util.InstanceClasses.AccountsRepoManager;
 import com.CSCI152.team4.server.Accounts.Requests.PermissionUpdateRequest;
+import com.CSCI152.team4.server.Accounts.Requests.TargetAccountRequest;
+import com.CSCI152.team4.server.Accounts.Requests.UpdateFromAdminRequest;
+import com.CSCI152.team4.server.Util.InstanceClasses.AccountsRepoManager;
 import com.CSCI152.team4.server.Util.InstanceClasses.Request;
 import com.CSCI152.team4.server.Util.InstanceClasses.TokenAuthenticator;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -12,26 +14,10 @@ import org.springframework.security.crypto.bcrypt.BCrypt;
 import org.springframework.stereotype.Service;
 import org.springframework.web.server.ResponseStatusException;
 
-import java.util.List;
 import java.util.function.Supplier;
 
 @Service
 public class AccountManagementService {
-
-    /**
-     * This Class Serves as a Getter/Setter Service
-     * for All account types.
-     *
-     * Post Requests will be handled by directly mapping
-     * the Supplied Account information to the database
-     * entities
-     *
-     * The exception to this rule is the 'Permissions'
-     * for an employee account. To change permissions,
-     * a PermissionChangeRequest will be used to ensure
-     * ONLY admins can change the permissions.
-     *
-     */
 
     private final AccountsRepoManager repos;
     private final TokenAuthenticator authenticator;
@@ -42,11 +28,80 @@ public class AccountManagementService {
         this.authenticator = authenticator;
     }
 
+    /*Get Own Account Info*/
+    public WorkerAccount getAccountInfo(Request request){
+        authenticator.validateToken(request.getToken(), request.getAccountIdString());
+
+        if(repos.businessExists(request.getBusinessId())){
+            WorkerAccount returnable
+                    = repos.getAccountByEmailAndBusinessId(request.getAccountEmail(),
+                    request.getBusinessId());
+            /*Do not return password!*/
+            returnable.setPassword(null);
+            return returnable;
+        }
+        throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "Business Account Not Found!");
+    }
+
+    /*Get Another Admins Account information: Only Available to Admin Accounts*/
+    public AdminAccount getOtherAdminAccountInfo(TargetAccountRequest request){
+        authenticator.validateToken(request.getToken(), request.getAccountIdString());
+        return (AdminAccount) getReturnableOnAdminExists(request.getAccountId(),
+                () -> repos.getAdminIfExists(request.getTargetId()));
+    }
+
+    /*Get Another Employees Account information: Only Available to Admin Accounts*/
+    public EmployeeAccount getOtherEmployeeAccountInfo(TargetAccountRequest request){
+        authenticator.validateToken(request.getToken(), request.getAccountIdString());
+        return (EmployeeAccount) getReturnableOnAdminExists(request.getAccountId(),
+                () -> repos.getEmployeeIfExists(request.getTargetId()));
+    }
+
+    /*An Admin will Update another Accounts information, but only the
+     * mutable fields
+     * Account Info is returned as proof
+     * of update*/
+    public WorkerAccount updateOtherFromAdmin(UpdateFromAdminRequest request){
+        authenticator.validateToken(request.getToken(), request.getAccountIdString());
+
+        return getReturnableOnAdminExists(request.getAccountId(),
+                () -> updateAndSaveOther(request));
+    }
+
+    /*Admin updates their own info, Account Info is returned as proof
+     * of update*/
     public AdminAccount updateAdminAccount(AdminAccount account){
         authenticator.validateToken(account.getToken(), account.getAccountIdString());
+        return updateAndSaveAdminAccount(account);
+    }
 
-        return (AdminAccount) getReturnableIfAdminExists(account.getAccountId(),
-                () -> updateAndSaveAdminAccount(account));
+    /*Employee Updates their own info, Account Info is returned as proof
+    * of update*/
+    public EmployeeAccount updateEmployeeAccount(EmployeeAccount account){
+        authenticator.validateToken(account.getToken(), account.getAccountIdString());
+        return updateAndSaveEmployeeAccount(account);
+    }
+
+    private WorkerAccount updateAndSaveOther(UpdateFromAdminRequest request){
+
+        BusinessAccount businessAccount = repos.getBusinessIfExists(request.getBusinessId());
+
+        String accountType = businessAccount.getAccountType(request.getTargetId().getAccountIdString());
+
+        switch(accountType){
+            case BusinessAccount.adminAccountType:
+                return updateAndSaveAdminAccount(
+                        new AdminAccount(request.getBusinessId(), request.getEmail(), null,
+                                request.getFirstName(), request.getLastName(), request.getJobTitle())
+                );
+            case BusinessAccount.employeeAccountType:
+                return updateAndSaveEmployeeAccount(
+                        new EmployeeAccount(request.getBusinessId(), request.getEmail(), null,
+                                request.getFirstName(), request.getLastName(), request.getJobTitle())
+                );
+            default:
+                throw new ResponseStatusException(HttpStatus.BAD_REQUEST,"Invalid Account Type!");
+        }
     }
 
     private AdminAccount updateAndSaveAdminAccount(AdminAccount account){
@@ -55,28 +110,11 @@ public class AccountManagementService {
 
         updateMutableAccountFields(account, accountToSave);
 
-        repos.saveAdminAccount(accountToSave);
+        AdminAccount returnable = repos.saveAdminAccount(accountToSave);
+        /*Do not return password!*/
+        returnable.setPassword(null);
 
-        return  getReturnableAdmin(account.getAccountId());
-    }
-
-    public ResponseEntity<Enum<HttpStatus>> updateEmployeeAccountFromAdmin(String token, AccountId adminAccountId, EmployeeAccount update){
-        authenticator.validateToken(token, adminAccountId.getAccountIdString());
-
-        if(repos.adminExists(adminAccountId)){
-            updateAndSaveEmployeeAccount(update);
-            return new ResponseEntity<>(HttpStatus.OK);
-        }
-
-        throw new ResponseStatusException(HttpStatus.BAD_REQUEST);
-    }
-
-    public EmployeeAccount updateEmployeeAccount(EmployeeAccount account){
-        authenticator.validateToken(account.getToken(), account.getAccountIdString());
-
-        return (EmployeeAccount) getReturnableIfEmployeeExists(account.getAccountId(),
-                () -> updateAndSaveEmployeeAccount(account));
-
+        return  returnable;
     }
 
     private EmployeeAccount updateAndSaveEmployeeAccount(EmployeeAccount account){
@@ -84,11 +122,17 @@ public class AccountManagementService {
 
         updateMutableAccountFields(account, accountToSave);
 
-        repos.saveEmployeeAccount(accountToSave);
+        EmployeeAccount returnable = repos.saveEmployeeAccount(accountToSave);
+        /*Do not return password!*/
+        returnable.setPassword(null);
 
-        return getReturnableEmployee(account.getAccountId());
+        return returnable;
     }
 
+    /*Non-mutable fields are Email, AccountId String and BusinessId
+    * All other fields are updatable
+    *
+    * Refer to WorkerAccount for available fields */
     private void updateMutableAccountFields(WorkerAccount newInfo, WorkerAccount originalInfo){
 
         String fNameToUpdate = newInfo.getFirstName();
@@ -114,90 +158,30 @@ public class AccountManagementService {
         }
     }
 
-    public WorkerAccount getAccountInfo(Request request){
-
-        if(repos.businessExists(request.getBusinessId())){
-            BusinessAccount businessAccount = repos.getBusinessIfExists(request.getBusinessId());
-
-            if(businessAccount.getAccountType(request.getAccountIdString())
-                    .equals(BusinessAccount.adminAccountType)){
-                return getAdminAccountInfo(request.getToken(), request.getAccountId());
-            } else if(businessAccount.getAccountType(request.getAccountIdString())
-                    .equals(BusinessAccount.employeeAccountType)) {
-                return getEmployeeAccountInfo(request.getToken(), request.getAccountId());
-            }
-        }
-
-        throw new ResponseStatusException(HttpStatus.BAD_REQUEST);
-
-    }
-
-    private AdminAccount getAdminAccountInfo(String token, AccountId accountId){
-        authenticator.validateToken(token, accountId.getAccountIdString());
-
-        return (AdminAccount) getReturnableIfAdminExists(accountId,
-                () -> getReturnableAdmin(accountId));
-    }
-
-    private EmployeeAccount getEmployeeAccountInfo(String token, AccountId accountId){
-        authenticator.validateToken(token, accountId.getAccountIdString());
-
-        return (EmployeeAccount) getReturnableIfEmployeeExists(accountId,
-                () -> getReturnableEmployee(accountId));
-    }
-
-    public AdminAccount getAdminAccountFromAdmin(String token, AccountId requestingAccount, AccountId targetAccount){
-        authenticator.validateToken(token, requestingAccount.getAccountIdString());
-        return (AdminAccount) getReturnableIfAdminExists(requestingAccount,
-                () -> getReturnableAdmin(targetAccount));
-    }
-
-    public EmployeeAccount getEmployeeAccountFromAdmin(String token, AccountId requestingAccount, AccountId targetAccount){
-        authenticator.validateToken(token, requestingAccount.getAccountIdString());
-
-        return (EmployeeAccount) getReturnableIfAdminExists(requestingAccount,
-                () -> getReturnableEmployee(targetAccount));
-    }
-
-    private AdminAccount getReturnableAdmin(AccountId accountId){
-        AdminAccount returnable = repos.getAdminIfExists(accountId);
-        if (returnable != null) returnable.setPassword("");
-        return returnable;
-    }
-
-    private EmployeeAccount getReturnableEmployee(AccountId accountId){
-        EmployeeAccount returnable = repos.getEmployeeIfExists(accountId);
-        if (returnable != null) returnable.setPassword(null);
-        return returnable;
-    }
-
-    public ResponseEntity<Enum<HttpStatus>> updateEmployeePermissions(PermissionUpdateRequest request){
+    /*Update an Employees Permissions from an Admin Account
+    * Return the Account permissions as proof of success*/
+    public WorkerAccount updateEmployeePermissions(PermissionUpdateRequest request){
         authenticator.validateToken(request.getToken(), request.getAccountIdString());
-
-        if(repos.adminExists(request.getAccountId())){
-            setEmployeePermissionsAndSave(request.getAccountToUpdateId(), request.getPermissions());
-            return new ResponseEntity<>(HttpStatus.OK);
-        }
-        return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+        return getReturnableOnAdminExists(request.getAccountId(),
+                () -> updatePermissions(request));
     }
 
-    private void setEmployeePermissionsAndSave(AccountId accountId, List<String> permissions){
-        EmployeeAccount accountToUpdate = repos.getEmployeeIfExists(accountId);
-        accountToUpdate.setPermissionsList(permissions);
-        repos.saveEmployeeAccount(accountToUpdate);
+    private WorkerAccount updatePermissions(PermissionUpdateRequest request){
+        EmployeeAccount accountToUpdate = repos.getEmployeeIfExists(request.getTargetId());
+        accountToUpdate.setPermissionsList(request.getPermissions());
+        return repos.saveEmployeeAccount(accountToUpdate);
     }
 
-    private WorkerAccount getReturnableIfAdminExists(AccountId adminAccountId, Supplier<WorkerAccount> toReturn){
-        if(repos.adminExists(adminAccountId)){
-            return toReturn.get();
+    /*Accepts a Lambda Expression to execute IF the admin is valid AND have matching business ID's*/
+    private WorkerAccount getReturnableOnAdminExists(AccountId requestingAccountId, Supplier<WorkerAccount> toReturn){
+        if(repos.adminExists(requestingAccountId)){
+            WorkerAccount returnable = toReturn.get();
+            if(returnable.getBusinessId().equals(requestingAccountId.getBusinessId())){
+                returnable.setPassword(null);
+                return returnable;
+            }
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "Invalid Business Id!");
         }
-        throw new ResponseStatusException(HttpStatus.BAD_REQUEST);
-    }
-
-    private WorkerAccount getReturnableIfEmployeeExists(AccountId employeeAccountId, Supplier<WorkerAccount> toReturn){
-        if(repos.employeeExists(employeeAccountId)){
-            return toReturn.get();
-        }
-        throw new ResponseStatusException(HttpStatus.BAD_REQUEST);
+        throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "Admin Does Not Exist!");
     }
 }
