+++ java services/server/server/src/main/java/com/CSCI152/team4/server/Accounts/Services/SessionService.java
@@ -5,6 +5,7 @@ import com.CSCI152.team4.server.Accounts.Classes.BusinessAccount;
 import com.CSCI152.team4.server.Accounts.Classes.WorkerAccount;
 import com.CSCI152.team4.server.Util.InstanceClasses.AccountsRepoManager;
 import com.CSCI152.team4.server.Accounts.Requests.LoginRequest;
+import com.CSCI152.team4.server.Util.InstanceClasses.Request;
 import com.CSCI152.team4.server.Util.InstanceClasses.TokenAuthenticator;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
@@ -68,9 +69,9 @@ public class SessionService {
 
     }
 
-    public ResponseEntity<Enum<HttpStatus>> logout(String token, AccountId accountId){
+    public ResponseEntity<Enum<HttpStatus>> logout(Request request){
 
-        authenticator.invalidateToken(token, accountId.getAccountIdString());
+        authenticator.invalidateToken(request.getToken(), request.getAccountIdString());
 
         return new ResponseEntity<Enum<HttpStatus>>(HttpStatus.OK);
     }
